Add private functions to displayController
Functions for checking if a player has won on a row, a column or a
diagonal. The functions return null if noone has won, and returns the
player that has won otherwise:
- _checkRow needs the row index from 0 to 2.
- _checkColumn needs the column index from 0 to 2.
- _checkDiagonal needs both the row and column index, from 0 to 2.

In these functions, row 0 corresponds to the top row and column 0
corresponds to the left column. When a new sign is added to the
gameBoard at position (row, column), only the relevant row, column and
diagonal has to be checked for a win using these functions.