Fixes switching player when clicking filled fields
Clicking a field would always activate the _switchPlayer and
_updatePlayerText functions in displayController, even if the field was
already filled. This happened because the check for a undefined (empty)
field in the gameBoard array (_board) was only done in the gameBoard
setField function.

To solve this, the check for a undefined field is instead done on every
card click event. The function getField from gameBoard is called and
functions are only called if the _board element is undefined.