Add win functionality
displayController has a private _winner variable which keeps track of
the winner player. If _winner is undefined, the game is still active and
a click of a card will still function.

Clicking a card will call _checkWinner(row, col) with the parameters:
- Math.floor(index / 3) for the row.
- index % 3 for the column.

In _checkWinner, both the row, column and diagonal is checked in order
to assign .winCard to all the cards that contribute to the win. The
_checkField function also assigns the winning player to _winner.

After _checkWinner is called two different situations can occur:
- There is a winner: _win() is called. This logs the a message of who
  won and adds the .game-over class to the gameBoard. This prevents the
  hover effects on cards after a winner has been decided.
- There is not a winner: _switchPlayer() is called and the game
  continues as usual.

reset() now also removes .game-over from gameBoard and sets _winner to
undefined when called.