Restructure main content
Adds divs for all the game content, the player-info and the restart
button. <main> has display: flex instead of grid to place the game
content in the middle of main. The game content then has a grid which
contains player-info, gameBoard and the restart button.

The JavaScript file is modified to adjust for the renaming of classes.