Implement player icons
Adds svg icons to index.html. The height of the player icons is the same
as the card height.

The current player will have a filled player icon and the other player
will have a player icon with only an outline. This is accomplished by
changing the content to a different url in the CSS file based on the
.current class.