Add transitions to player icons
To allow transitions, the icons needed to be changed using background-
image instead of using content. Having the no-repeat property in the
background shorthand and contain for background-size gives the icon the
corrent size.

The switch to background-image also means .player-icon does not have to
be a img element. It is therefore changed to a div in index.html.